Use Dependabot to check for outdated action versions used in workflows
Dependabot will periodically check the GitHub Actions workflows and workflow templates of the repository and submit pull
requests to update any that are found to be using an outdated version of an action.

This check has previously been a manual process, and a frequently neglected one. Up to date action versions are
especially important in the workflow templates, since they are the source for the CI systems of many repositories.

Typically, the best approach is to only pin the major version of the action (e.g., `uses: foo/bar@v1`), so the workflow
will automatically use the latest version of the action except when there have been breaking changes. Many actions
provide major version refs for this purpose. Dependabot's PRs may propose to use a ref that pins the patch version
(e.g., `uses: foo/bar@v2.3.4`). In this case, when a major version ref is available, the Dependabot PR should be treated
as only an update notification, and the update done manually, at which time the Dependabot PR will be closed
automatically.

In the case of PRs from Dependabot for updates to the workflow template copies, these serve only as update notifications
Dependabot has been configured to prefix the PR message with "(DO NOT MERGE)" to make this clear.