Remove donation links
The Arduino company is no longer soliciting monetary donations. For this reason, the donation link is hereby removed
from the organization profile, and from the "Sponsor this project" section of each repository homepage (as was
previously configured via the FUNDING.yml file).

Community members who wish to contribute to the project still have opportunities to do so via the other options listed
in the organization profile.